Refactor state structure
