+++ plugins/rich-editor/src/scripts/components/editor/Editor.tsx
@@ -19,10 +19,10 @@ import UploadButton from "@rich-editor/components/editor/pieces/EditorUploadButt
 import { EditorProvider } from "@rich-editor/components/context";
 import EditorDescriptions from "@rich-editor/components/editor/pieces/EditorDescriptions";
 import { Provider as ReduxProvider } from "react-redux";
-import IState from "@rich-editor/state/IState";
 import { actions } from "@rich-editor/state/instance/instanceActions";
 import { getIDForQuill, isEmbedSelected, SELECTION_UPDATE } from "@rich-editor/quill/utility";
 import { FOCUS_CLASS } from "@dashboard/embeds";
+import { IStoreState } from "@rich-editor/@types/store";
 
 interface IProps {
     editorID: string;
@@ -33,7 +33,7 @@ interface IProps {
 export default class Editor extends React.Component<IProps> {
     private hasUploadPermission: boolean;
     private quillMountRef: React.RefObject<HTMLDivElement> = React.createRef();
-    private store = getStore<IState>();
+    private store = getStore<IStoreState>();
     private allowPasteListener = true;
     private editorID: string;
     private quill: Quill;
