+++ plugins/rich-editor/src/scripts/quill/ClipboardModule.ts
@@ -10,8 +10,8 @@ import Quill, { DeltaStatic } from "quill/core";
 import { rangeContainsBlot, getIDForQuill } from "@rich-editor/quill/utility";
 import CodeBlockBlot from "@rich-editor/quill/blots/blocks/CodeBlockBlot";
 import CodeBlot from "@rich-editor/quill/blots/inline/CodeBlot";
-import IState from "@rich-editor/state/IState";
 import getStore from "@dashboard/state/getStore";
+import { IStoreState } from "@rich-editor/@types/store";
 
 export default class ClipboardModule extends ClipboardBase {
     /**
@@ -106,7 +106,7 @@ export default class ClipboardModule extends ClipboardBase {
      * Determine if we are in a code formatted item or not.
      */
     private get inCodeFormat() {
-        const instance = getStore<IState>().getState().editor.instances[getIDForQuill(this.quill)];
+        const instance = getStore<IStoreState>().getState().editor.instances[getIDForQuill(this.quill)];
         if (!instance || !instance.lastGoodSelection) {
             return false;
         }
