+++ plugins/rich-editor/src/scripts/quill/EmbedInsertionModule.ts
@@ -11,14 +11,14 @@ import api, { uploadImage } from "@dashboard/apiv2";
 import { getPastedImage, getDraggedImage } from "@dashboard/dom";
 import ExternalEmbedBlot, { IEmbedValue } from "./blots/embeds/ExternalEmbedBlot";
 import getStore from "@dashboard/state/getStore";
-import IState from "@rich-editor/state/IState";
 import { getIDForQuill, insertBlockBlotAt } from "@rich-editor/quill/utility";
+import { IStoreState } from "@rich-editor/@types/store";
 
 /**
  * A Quill module for managing insertion of embeds/loading/error states.
  */
 export default class EmbedInsertionModule extends Module {
-    private store = getStore<IState>();
+    private store = getStore<IStoreState>();
 
     constructor(public quill: Quill, options = {}, editorID) {
         super(quill, options);
