+++ plugins/rich-editor/src/scripts/state/mention/mentionReducer.ts
@@ -7,7 +7,7 @@
 import * as mentionActions from "./mentionActions";
 import MentionTrie from "@rich-editor/state/mention/MentionTrie";
 import { IMentionSuggestionData } from "@rich-editor/components/toolbars/pieces/MentionSuggestion";
-import { IMentionState } from "@rich-editor/state/IState";
+import { IMentionState } from "@rich-editor/@types/store";
 
 export const initialState: IMentionState = {
     lastSuccessfulUsername: null,
